Move RemoveEmptyStrings to stringhelper
