+++ pkg/stringhelper/stringhelper.go
@@ -48,3 +48,13 @@ func RemoveDuplicates(str []string) []string {
 	}
 	return list
 }
+
+func RemoveEmptyStrings(s []string) []string {
+	var r []string
+	for _, str := range s {
+		if str != "" {
+			r = append(r, str)
+		}
+	}
+	return r
+}
